content-repo-creator/requirements.txt
github3.py==2.0.0
GitPython==3.1.14